Convert RoomReleaseGroupArtistCredit to 1-to-many from many-to-many
A release group will have 1+ artist credit(s). However, the same artist can be listed multiple times, so this is not a many-to-many between release groups and artists.
We can already find all release groups by artists through queries, so this doesn't serve that purpose.
The only purpose of this table is to store all artist credits for a release group.